ai-service/reddit_test/requirements.txt
praw
python-dotenv
mistralai
